refactor: restructure tests and clean up root directory
- Move loose tests to agent/, cli/, core/, and tools/

- Move scripts to scripts/

- Move tasks.md to docs/

- Add CHANGES.md tracking file migrations
