+++ tests/cli/test_cli_main.py
