+++ tests/cli/test_dashboard_smoke.py
