+++ tests/core/test_config_manager.py
