+++ tests/core/test_history_manager.py
