+++ tests/core/test_identity_manager.py
