+++ tests/core/test_memory_manager.py
