+++ tests/core/test_metrics.py
