+++ tests/core/test_mission_manager.py
