+++ tests/core/test_paths.py
