+++ tests/core/test_tasks_manager.py
