+++ tests/tools/test_memory_tools.py
